Updating the id values of <i>
So when I added pauseall() and startplay() to the code which is responsible for changing the pause and play button in front of each song, it showed an error. The error arise due the fact that the numbering in the <i> id was not proper which was giving a null value which caused the functions (pauseall() and startplay()) to not work.